requirements.txt
Flask
markdown
python-frontmatter
PyGithub
requests
python-dotenv
supabase
